Update README.md
POLICE DEATH IN AMERICA
This project contains SQL queries written to query Police death in America for key insights to help detect region 
with the highest crime rate and the ranks affected the most. This dataset contains information on fatal police deaths 
in the United States. The data includes the victim's rank, name, department, date of death, and cause of death. 
The data spans from 1791 to the present day.
Data Source:
The data was gotten from kaggle.com and it contains10 variables. The dataset contains information of police like
 Rank (String) -> Rank assigned or achieved by the police throughout their tenure.
Name (String) -> The name of the person
CauseofDeath (String) -> The cause of the death
Date (String) -> The full date of the incident
Year (Int) -> The year of the incident
Month (String) -> The month of the incident
Day (String) -> The day of the incident
Department (String)-> The department's name and the person who works there.
State (String) -> The state where the department is situated
K9_Unit (Boolean)-> K9, the nickname of police dogs and the police dog unit itself